:recycle: Refactor BinarySearch algorithm and separate responsibilities
